compute indicators with SQL queries
git-svn-id: http://talendforge.org/svn/tis_shared/trunk@13810 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.cwm.management/src/test/java/org/talend/dq/indicators/LoadSerialData.java
@@ -13,6 +13,7 @@
 package org.talend.dq.indicators;
 
 import java.io.File;
+import java.math.BigInteger;
 
 import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.common.util.URI;
@@ -57,7 +58,7 @@ public class LoadSerialData {
         for (EObject object : contents) {
             FrequencyIndicator freqI = mySwitch.doSwitch(object);
             if (freqI != null) {
-                int uniqueValueCount = freqI.getUniqueValueCount();
+                BigInteger uniqueValueCount = freqI.getUniqueValueCount();
                 System.out.println("nb unique values = " + uniqueValueCount);
                 EList<Object> uniqueValues = freqI.getUniqueValues();
                 for (Object data : uniqueValues) {
